+++ imperatives/mocktable_test.go
@@ -12,7 +12,7 @@ type mockTable struct {
 
 func (m *mockTable) AddAggregator(agg *aggregator.Aggregator) {}
 func (m *mockTable) AddRewriter(rw rewriter.RW)               {}
-func (m *mockTable) AddBlacklist(matcher *matcher.Matcher)    {}
+func (m *mockTable) AddBlocklist(matcher *matcher.Matcher)    {}
 func (m *mockTable) AddRoute(route route.Route)               {}
 func (m *mockTable) DelRoute(key string) error                { return nil }
 func (m *mockTable) UpdateDestination(key string, index int, opts map[string]string) error {